requirements.txt
Flask==0.10.1
Werkzeug==0.10.1
argparse==1.2.1
pyOpenSSL==0.13
requests==2.2.1
